requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
beautifulsoup4==4.7.1
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
future==0.17.1
pycodestyle==2.5.0
pycparser==2.19
python-telegram-bot==12.0.0b1
six==1.12.0
soupsieve==1.9.2
tornado==6.0.3
urllib3==1.25.3
